requirements.txt
altair==5.1.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
distlib==0.3.7
filelock==3.12.4
gitdb==4.0.10
GitPython==3.1.36
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.25.2
packaging==23.1
pandas==2.1.0
Pillow==9.5.0
pipenv==2023.9.8
platformdirs==3.10.0
protobuf==4.24.3
pyarrow==13.0.0
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
python-dateutil==2.8.2
python-pptx==0.6.22
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.10.3
six==1.16.0
smmap==5.0.0
streamlit==1.26.0
streamlit-pretty-notification-box==0.0.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.22.0
virtualenv==20.24.5
XlsxWriter==3.1.3
zipp==3.16.2
